New noise stuff for spectralacim functions
Added CONpts option to specacim
Added lastspiketonoise option to specacim to make the transfer of the
last spike in critical orbits an evenly distributed function
Allowed more than one whichsp and whichnsp to be passed to spikefn
Added fastinv options to spectralacimpf, allowing the invariant measure
to be obtained from the transfer operator in O(N^2) time, possibly at
the expense of accuracy.
In legconvgetptmatrix, truncated different terms in the recurrence relation at the
matrix edge to stop it blowing up at the edges
Added leghconv, which does a mirrored convolution with a function on
[0,2] instead of [-1,1]: [-1,0] maps to [2,1] (orientation reversed) to
get the same thing. No transforming to value space and back required.
Added leggaushcoefs which finds coefficients for the Gaussian for
leghconv
Added separate fourier/leggauskernel functions
Changed leggauskernel to use hconv